WIP: MDEV-23538: Rename mariadb.pc to mariadbd.pc to avoid confusion
Change so that the client library config can be accessed using the name
libmariadb.pc or mariadb.pc. Rename old mariadb.pc to mariadbd.pc to better
indicate that the config is for the daemon build, not client.

In downstream Debian we already ship the libmariadb.pc as mariadb.pc:
https://salsa.debian.org/mariadb-team/mariadb-10.5/-/commit/2f183af990fbe1cfa8c343998c7640f45f45368b